Add Service.swift and implement fetchData method
This commit adds a new file called Service.swift to the WeatherApp project. The Service struct contains properties for latitude, longitude, and name of a city. It also includes a class called Service with a fetchData method that takes in a City object and performs an API request using URLSession.shared. The completion handler returns the string "Oi".

In ViewController.swift, an instance of the Service class is created and the fetchData method is called with sample data for Belo Horizonte.